Use the real Allegra and Mary era tags
Fixes #2668.

Previously, the Allegra and Mary eras were exactly the same as the Shelley era,
so no translations between the eras were needed.

In this commit, we use the actual, different `AllegraEra` and `ShelleyEra` tags.
This means we need translations between the eras.

In the `db-analyser`, split up `HasAnalysis` in `HasAnalysis` and
`HasProtocolInfo`. The Allegra and Mary eras have `HasAnalysis` instances, but
not yet `HasProtocolInfo` (and Allegra might never need one).